32-bit angles in minutiae and edges

+++ src/main/java/com/machinezoo/sourceafis/engine/features/FeatureMinutia.java
@@ -5,9 +5,9 @@ import com.machinezoo.sourceafis.engine.primitives.*;
 
 public class FeatureMinutia {
 	public final IntPoint position;
-	public final double direction;
+	public final float direction;
 	public final MinutiaType type;
-	public FeatureMinutia(IntPoint position, double direction, MinutiaType type) {
+	public FeatureMinutia(IntPoint position, float direction, MinutiaType type) {
 		this.position = position;
 		this.direction = direction;
 		this.type = type;
